64bit: start using stdint in syscall implementation

+++ Source/OS9exec_core/fileaccess.c
@@ -196,41 +196,41 @@
 void   init_File(fmgr_typ *f);
 os9err pFopen(ushort pid, syspath_typ *, ushort *modeP, const char *pathname);
 os9err pFclose(ushort pid, syspath_typ *);
-os9err pFread(ushort pid, syspath_typ *, ulong *n, char *buffer);
-os9err pFreadln(ushort pid, syspath_typ *, ulong *n, char *buffer);
-os9err pFwrite(ushort pid, syspath_typ *, ulong *n, char *buffer);
-os9err pFwriteln(ushort pid, syspath_typ *, ulong *n, char *buffer);
-os9err pFseek(ushort pid, syspath_typ *, ulong *posP);
+os9err pFread(ushort pid, syspath_typ *, uint32_t *n, char *buffer);
+os9err pFreadln(ushort pid, syspath_typ *, uint32_t *n, char *buffer);
+os9err pFwrite(ushort pid, syspath_typ *, uint32_t *n, char *buffer);
+os9err pFwriteln(ushort pid, syspath_typ *, uint32_t *n, char *buffer);
+os9err pFseek(ushort pid, syspath_typ *, uint32_t *posP);
 os9err pFdelete(ushort pid, syspath_typ *, ushort *modeP, char *pathname);
 
-os9err pFsize(ushort pid, syspath_typ *, ulong *sizeP);
+os9err pFsize(ushort pid, syspath_typ *, uint32_t *sizeP);
 os9err pFopt(ushort pid, syspath_typ *, byte *buffer);
 os9err pHvolnam(ushort pid, syspath_typ *, char *volname);
-os9err pFpos(ushort pid, syspath_typ *, ulong *posP);
+os9err pFpos(ushort pid, syspath_typ *, uint32_t *posP);
 os9err pFeof(ushort pid, syspath_typ *);
-os9err pFready(ushort pid, syspath_typ *, ulong *n);
-os9err pHgetFD(ushort pid, syspath_typ *, ulong *maxbytP, byte *buffer);
+os9err pFready(ushort pid, syspath_typ *, uint32_t *n);
+os9err pHgetFD(ushort pid, syspath_typ *, uint32_t *maxbytP, byte *buffer);
 os9err pHgetFDInf(ushort pid,
                   syspath_typ *,
                   ulong *maxbytP,
                   ulong *fdinf,
                   byte  *buffer);
 
-os9err pFsetsz(ushort pid, syspath_typ *, ulong *sizeP);
+os9err pFsetsz(ushort pid, syspath_typ *, uint32_t *sizeP);
 os9err pHsetFD(ushort pid, syspath_typ *, byte *buffer);
 
 void   init_Dir(fmgr_typ *f);
 os9err pDopen(ushort pid, syspath_typ *, ushort *modeP, const char *pathname);
 os9err pDclose(ushort pid, syspath_typ *);
-os9err pDread(ushort pid, syspath_typ *, ulong *n, char *buffer);
-os9err pDseek(ushort pid, syspath_typ *, ulong *posP);
+os9err pDread(ushort pid, syspath_typ *, uint32_t *n, char *buffer);
+os9err pDseek(ushort pid, syspath_typ *, uint32_t *posP);
 os9err pDchd(ushort pid, syspath_typ *, ushort *modeP, char *pathname);
 os9err pDmakdir(ushort pid, syspath_typ *, ushort *modeP, char *pathname);
 
-os9err pDsize(ushort pid, syspath_typ *, ulong *sizeP);
-os9err pDpos(ushort pid, syspath_typ *, ulong *posP);
+os9err pDsize(ushort pid, syspath_typ *, uint32_t *sizeP);
+os9err pDpos(ushort pid, syspath_typ *, uint32_t *posP);
 os9err pDeof(ushort pid, syspath_typ *);
-os9err pDsetatt(ushort pid, syspath_typ *, ulong *attr);
+os9err pDsetatt(ushort pid, syspath_typ *, uint32_t *attr);
 /* ------------------------------------------------------------------------- */
 
 void init_File(fmgr_typ *f)
@@ -308,7 +308,7 @@ void init_Dir(fmgr_typ *f)
 
 
 /* input from file */
-os9err pFread(_pid_, syspath_typ *spP, ulong *n, char *buffer)
+os9err pFread(_pid_, syspath_typ *spP, uint32_t *n, char *buffer)
 {
     long      cnt, k;
     file_typ *f = &spP->u.disk.u.file;
@@ -353,7 +353,7 @@ os9err pFread(_pid_, syspath_typ *spP, ulong *n, char *buffer)
 }
 
 /* input line from file */
-os9err pFreadln(_pid_, syspath_typ *spP, ulong *n, char *buffer)
+os9err pFreadln(_pid_, syspath_typ *spP, uint32_t *n, char *buffer)
 {
     long      cnt;
     file_typ *f = &spP->u.disk.u.file;
@@ -394,7 +394,7 @@ os9err pFreadln(_pid_, syspath_typ *spP, ulong *n, char *buffer)
 }
 
 /* output to file */
-os9err pFwrite(_pid_, syspath_typ *spP, ulong *n, char *buffer)
+os9err pFwrite(_pid_, syspath_typ *spP, uint32_t *n, char *buffer)
 {
     file_typ *f = &spP->u.disk.u.file;
 
@@ -428,7 +428,7 @@ os9err pFwrite(_pid_, syspath_typ *spP, ulong *n, char *buffer)
 }
 
 /* output to file */
-os9err pFwriteln(_pid_, syspath_typ *spP, ulong *n, char *buffer)
+os9err pFwriteln(_pid_, syspath_typ *spP, uint32_t *n, char *buffer)
 {
     file_typ *f = &spP->u.disk.u.file;
 
@@ -479,7 +479,7 @@ os9err pFopt(ushort pid, syspath_typ *spP, byte *buffer)
 }
 
 /* check ready */
-os9err pFready(_pid_, _spP_, ulong *n)
+os9err pFready(_pid_, _spP_, uint32_t *n)
 {
     *n = 1;
     return 0;
@@ -687,7 +687,7 @@ os9err pFclose(_pid_, syspath_typ *spP)
     return 0;
 }
 
-os9err pFseek(_pid_, syspath_typ *spP, ulong *posP)
+os9err pFseek(_pid_, syspath_typ *spP, uint32_t *posP)
 /* seek within a file */
 {
     int fildes;
@@ -764,15 +764,15 @@ os9err pFdelete(ushort pid, _spP_, ushort *modeP, char *pathname)
 }
 
 /* get file position */
-os9err pFpos(_pid_, syspath_typ *spP, ulong *posP)
+os9err pFpos(_pid_, syspath_typ *spP, uint32_t *posP)
 {
-    *posP = (ulong)ftell(spP->stream);
+    *posP = (uint32_t)ftell(spP->stream);
     // fgetpos( spP->stream,  posP );   /* save current position */
     return 0;
 }
 
 /* get file size */
-os9err pFsize(_pid_, syspath_typ *spP, ulong *sizeP)
+os9err pFsize(_pid_, syspath_typ *spP, uint32_t *sizeP)
 {
     os9err err = 0;
 
@@ -788,7 +788,7 @@ os9err pFsize(_pid_, syspath_typ *spP, ulong *sizeP)
 }
 
 /* set file size */
-os9err pFsetsz(ushort pid, syspath_typ *spP, ulong *sizeP)
+os9err pFsetsz(ushort pid, syspath_typ *spP, uint32_t *sizeP)
 {
     os9err err = 0;
     char   b   = 0;
@@ -1144,7 +1144,7 @@ static void setFD(syspath_typ *spP, void *fdl, byte *buffer)
 }
 
 /* get file descriptor for object */
-os9err pHgetFD(_pid_, syspath_typ *spP, ulong *maxbytP, byte *buffer)
+os9err pHgetFD(_pid_, syspath_typ *spP, uint32_t *maxbytP, byte *buffer)
 {
     void *fdl;
 
@@ -1271,7 +1271,7 @@ os9err pDclose(_pid_, syspath_typ *spP)
 }
 
 /* read from (simulated) directory file */
-os9err pDread(_pid_, syspath_typ *spP, ulong *n, char *buffer)
+os9err pDread(_pid_, syspath_typ *spP, uint32_t *n, char *buffer)
 {
     os9err err;
 
@@ -1370,14 +1370,14 @@ os9err pDread(_pid_, syspath_typ *spP, ulong *n, char *buffer)
 }
 
 /* get pointer position */
-os9err pDpos(_pid_, syspath_typ *spP, ulong *posP)
+os9err pDpos(_pid_, syspath_typ *spP, uint32_t *posP)
 {
     *posP = spP->u.disk.u.dir.pos;
     return 0;
 }
 
 /* get size of directory file */
-os9err pDsize(_pid_, syspath_typ *spP, ulong *sizeP)
+os9err pDsize(_pid_, syspath_typ *spP, uint32_t *sizeP)
 {
     os9err err = 0;
 
@@ -1387,7 +1387,7 @@ os9err pDsize(_pid_, syspath_typ *spP, ulong *sizeP)
 }
 
 /* set read position */
-os9err pDseek(ushort pid, syspath_typ *spP, ulong *posP)
+os9err pDseek(ushort pid, syspath_typ *spP, uint32_t *posP)
 {
     ulong       cnt;
     int         n;
@@ -1495,7 +1495,7 @@ os9err pDmakdir(ushort pid, _spP_, ushort *modeP, char *pathname)
     return host2os9err(oserr, E_WRITE);
 }
 
-os9err pDsetatt(ushort pid, syspath_typ *spP, ulong *attr)
+os9err pDsetatt(ushort pid, syspath_typ *spP, uint32_t *attr)
 {
     os9err err = 0;
